move ConfigHelper to Core folder do code cleanup add a ConfigPathHelper.java
